+++ util/src/shared/isNode.ts
@@ -1,3 +1,0 @@
-export const isNode = () => {
-    return typeof process === 'object';
-};