+++ util/CHANGELOG.md
@@ -5,7 +5,7 @@ Refer to https://keepachangelog.com/en/1.0.0/ for how to maintain changelog<br /
 
 ## [Unreleased]
 
--
+-   added Logger class
 
 ## [0.2.1] - 10/03/2023